// src/types/email.ts
export type IConfirmEmailTicket = {
  toName: string;
  toEmail: string;
  eventName: string;
  startDate: string;
  startTime: string;
  location: string;
  category: string;
  fromName: string;
  fromEmail: string;
  endTime: string;
  creatorName: string;
  creatorEmail: string;
};

export type ISendPaymentInfo = {};